Make each shader program into a singleton
