+++ src/BillboardCollection.ts
@@ -3,9 +3,11 @@ import { BillboardSpec, TypedArray, WebGLAnyRenderingContext } from "./AutumnTyp
 import { getGLFormatType } from "./PlotComponent";
 import { RawVectorField } from "./RawField";
 import { WGLBuffer, WGLProgram, WGLTexture, WGLTextureSpec } from "autumn-wgl";
+import { Cache } from "./utils";
 
 const billboard_vertex_shader_src = require('./glsl/billboard_vertex.glsl');
 const billboard_fragment_shader_src = require('./glsl/billboard_fragment.glsl');
+const program_cache = new Cache((gl: WebGLAnyRenderingContext) => new WGLProgram(gl, billboard_vertex_shader_src, billboard_fragment_shader_src))
 
 class BillboardCollection<ArrayType extends TypedArray> {
     readonly spec: BillboardSpec;
@@ -26,7 +28,7 @@ class BillboardCollection<ArrayType extends TypedArray> {
         this.color = billboard_color;
         this.size_multiplier = billboard_size_mult;
 
-        this.program = new WGLProgram(gl, billboard_vertex_shader_src, billboard_fragment_shader_src);
+        this.program = program_cache.getValue(gl);
         this.vertices = null;
         this.texcoords = null;
